优化代码风格，移除多余的修饰符和注释

+++ src/main/java/me/aleksilassila/litematica/printer/printer/zxy/Utils/overwrite/MyBox.java
@@ -19,7 +19,7 @@ public class MyBox extends Box implements Iterable<BlockPos> {
     }
 
     public MyBox(BlockPos pos) {
-        this((double) pos.getX(), (double) pos.getY(), (double) pos.getZ(), (double) (pos.getX()), (double) (pos.getY()), (double) (pos.getZ()));
+        this(pos.getX(), pos.getY(), pos.getZ(), pos.getX(), pos.getY(), pos.getZ());
     }
 
     public MyBox(Vec3d pos1, Vec3d pos2) {
